Call route callback even when route assigned to module
When routes are assigned to child modules, then they are automatically
loaded when the module router triggers the route event. This is fine in
most cases. But when the module needs to perform tasks before the child
is loaded, this commit adds the possibility to define a function with
the route name which will be called instead of autoloading the child module.

The child module can be manually loaded with the loadModule function of
the parent.